move helpers to helpers.js
