+++ plugins/Files/js/sagas/files.js
@@ -3,26 +3,7 @@ import { put } from 'redux-saga/effects'
 import * as actions from '../actions/files.js'
 import * as constants from '../constants/files.js'
 import BigNumber from 'bignumber.js'
-import { List } from 'immutable'
-
-const sendError = (e) => {
-	SiaAPI.showError({
-		title: 'Sia-UI Files Error',
-		content: e.toString(),
-	})
-}
-
-// siadCall: promisify Siad API calls.  Resolve the promise with `response` if the call was successful,
-// otherwise reject the promise with `err`.
-const siadCall = (uri) => new Promise((resolve, reject) => {
-	SiaAPI.call(uri, (err, response) => {
-		if (err) {
-			reject(err)
-		} else {
-			resolve(response)
-		}
-	})
-})
+import { sendError, siadCall, parseFiles } from './helpers.js'
 
 // Query siad for the state of the wallet.
 // dispatch `unlocked` in receiveWalletLockstate
@@ -33,16 +14,6 @@ function* getWalletLockstateSaga() {
 	} catch (e) {
 		sendError(e)
 	}
-}
-
-// Parse the response from `/renter/files`
-const parseFiles = (files) => {
-	const fileList = List(files)
-	return fileList.map((file) => ({
-		size: file.filesize,
-		name: file.siapath,
-		available: file.available,
-	}))
 }
 
 // Query siad for the user's files.
